+++ src/android_linux/asm_generic/sockios.rs
@@ -2,13 +2,13 @@
 // Copyright © 2016 The developers of libc-extra. See the COPYRIGHT file in the top-level directory of this distribution and at https://raw.githubusercontent.com/lemonrock/libc-extra/master/COPYRIGHT.
 
 
-use ::libc::c_ulong;
+use ::libc::c_int;
 
 
-pub const FIOSETOWN: c_ulong = 0x8901;
-pub const SIOCSPGRP: c_ulong = 0x8902;
-pub const FIOGETOWN: c_ulong = 0x8903;
-pub const SIOCGPGRP: c_ulong = 0x8904;
-pub const SIOCATMARK: c_ulong = 0x8905;
-pub const SIOCGSTAMP: c_ulong = 0x8906;
-pub const SIOCGSTAMPNS: c_ulong = 0x8907;
+pub const FIOSETOWN: c_int = 0x8901;
+pub const SIOCSPGRP: c_int = 0x8902;
+pub const FIOGETOWN: c_int = 0x8903;
+pub const SIOCGPGRP: c_int = 0x8904;
+pub const SIOCATMARK: c_int = 0x8905;
+pub const SIOCGSTAMP: c_int = 0x8906;
+pub const SIOCGSTAMPNS: c_int = 0x8907;
